pyproject.toml
[project]
name = "ai-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.77.0",
    "edgartools>=5.30.0",
    "groq>=1.0.0",
    "hdbscan>=0.8.42",
    "langextract>=1.1.1",
    "nltk>=3.9.3",
    "numpy>=2.4.2",
    "qdrant-client[fastembed]>=1.17.0",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.3",
    "whoosh>=2.7.4",
]
